Don't deliver responses out of band
This change ensures that responses and notifications are dispatched in
the order they are received with respect to each other. This is
important especially for pubsub, where the response to the initial
subscription request carries the subscription ID, knowledge of which
is required to dispatch later subscription updates to the correct
recipient.